requirements.txt
Django==6.0.3
djangorestframework==3.17.0
django-filter==25.2
python-decouple==3.8
